Simulate Gaussian noise from detector PSD

Generate colored Gaussian noise with power spectral density of a given detector.

Usage

psd_noise(
  det,
  duration,
  tstart,
  sampling.freq = 4096,
  fl = 15,
  delta_f = 1/32,
  seed = NULL
)

Arguments

det

A character. Detector abbreviation ("H1", "L1", "V1", "K1", or "E1").

duration

A numeric. Noise duration (in seconds).

tstart

A numeric. Start GPS time of the noise.

sampling.freq

A numeric. Sampling frequency (Hz). Default: 4096.

fl

A numeric. Lower frequency cutoff. Default: 15 Hz.

delta_f

A numeric. Frequency resolution. Default: 1/32.

seed

Optional numeric. Random seed for reproducibility.

Value

A ts object of simulated Gaussian noise.

Details

This function interfaces with Python's pycbc library using reticulate. Please ensure that the required Python packages are available in your environment.