Save a Data Frame as a Markdown-Formatted Table

Save a data frame in markdown format using insight::export_table.

Usage

savetab(
  x,
  file,
  signif = 3,
  col.names = F,
  row.names = F,
  quote = F,
  show = T,
  append = F,
  nsep = 30,
  ...
)

Arguments

x

A data.frame to save.

file

A character string indicating the file path.

signif

Integer; number of significant digits (default: 3).

col.names

Logical; whether to include column names (default: FALSE).

row.names

Logical; whether to include row names (default: FALSE).

quote

Logical; whether to quote each value (default: FALSE).

show

Logical; if TRUE, print the table content (default: TRUE).

append

Logical; if TRUE, append to the file (default: FALSE).

nsep

Integer; number of dashes used to separate appended tables (default: 30).

...

Additional arguments for insight::export_table.

Value

None.